requirements.txt
django
djangorestframework
pytest
pytest-django
requests
# django-cors-headers